Handle cells that go over the edge of the image
It is possible for the values in the bounding matrices (X and Y) to be less than 1 or greater than the maximum dimension of the image matrix. These can then be rounded down to 0 or up too high such that an out of index error is thrown. This commit changes the behavior to default the bounding box to the edge of the frame if it would otherwise go over.